move registrable function
